// ionic/wj/src/app/interfaces/TastingSystemNose.ts
export interface TastingSystemNose {
  name: string;
  intensities: string[];
  intensityTitle: string;
  floral: string[];
  greenFruit: string[];
  citrusFruit: string[];
  ripeness: string[];   
  stoneFruit: string[];
  tropicalFruit: string[];
  redFruit: string[];
  blackFruit: string[];
  herbaceous: string[];
  herbal: string[];
  spice: string[];

}